backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone
pillow
python-multipart
requests
PyPDF2
pymupdf
youtube-transcript-api
